requirements.txt
annotated-types==0.5.0
anyio==3.7.1
asyncpg==0.28.0
boto3==1.28.32
botocore==1.31.32
cffi==1.15.1
click==8.1.7
cryptography==41.0.3
dnspython==2.4.2
ecdsa==0.18.0
fastapi==0.101.1
h11==0.14.0
httptools==0.6.0
idna==3.4
jmespath==1.0.1
passlib==1.7.4
prettytable==3.9.0
psycopg2-binary==2.9.7
pyasn1==0.5.0
pycparser==2.21
pydantic==2.2.1
pydantic_core==2.6.1
pymongo==4.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
rsa==4.9
s3transfer==0.6.2
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
urllib3==1.26.16
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.3
